requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
bleach==4.1.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
docutils==0.18.1
ghp-import==2.0.2
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
Jinja2==3.0.3
keyring==23.5.0
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.3
outcome==1.1.0
packaging==21.3
Pillow==9.0.0
pkginfo==1.8.2
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0
pyyaml_env_tag==0.1
readme-renderer==32.0
regex==2021.11.10
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
selenium==4.12.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
twine==3.7.1
urllib3==1.26.8
watchdog==2.1.6
webencodings==0.5.1
wsproto==1.0.0
zipp==3.7.0
